sdcard_sd: Decode commands always as command
The SD card decoder predicts always the next token type from the
current token type. This is working if the SD card is responding as
expected.
But if the card is not responding, the prediction is wrong. A second
command token is decoded as a response token.

To fix this behavior the decoder state is set to St.GET_COMMAND_TOKEN,
if the current token is a command token.