sdcard_sd: Redesign state machine
The decoder reads a complete token before a token handler is called.
If the token is complete, the common token fields are handled.
Depending on the Transmission bit the corresponding handler is called.
For a command, the handler is determined by the Command field. The
command handler specifies the next expected response type and the
expected token length.
For a response, the previously specified handler is called.